+++ src/module/car/module.js
@@ -1,6 +1,7 @@
+require('./types/car.dto')
 const CarController = require('./controller/car.controller')
 const CarService = require('./car.service')
-const CarRepository = require('./repository/sqlite/carRepository')
+const CarRepository = require('./repository/sqlite/car.repository')
 
 /**
  * @param {import('express').Application} app
