+++ src/module/car/types/car.dto.js
@@ -32,5 +32,5 @@
  *  automatic: (1|0),
  *  price_per_week_in_dollars: number,
  *  price_per_day_in_dollars: number
- * }} CarFromHttpDto
+ * }} CarFromHttpRequestDto
 */
